+++ src/Models/MailchimpList.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Ryssbowh\CraftEmails\Models;
+
+use yii\base\DynamicModel;
+
+class MailchimpList extends DynamicModel
+{
+    /**
+     * @var MailchimpMember[]
+     */
+    public $members = [];
+
+    /**
+     * Get all subscribed members emails of this list
+     * 
+     * @return string[]
+     */
+    public function getEmails(): array
+    {
+        $emails = [];
+        foreach ($this->members as $member) {
+            if ($member->status == 'subscribed') {
+                $emails[$member->email_address] = $member->full_name;
+            }
+        }
+        return $emails;
+    }
+}