+++ src/Models/MailchimpMember.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Ryssbowh\CraftEmails\Models;
+
+use yii\base\DynamicModel;
+
+class MailchimpMember extends DynamicModel
+{
+}