+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # ryssbowh/craft-emails Changelog
 
+## 1.0.1 - 09/01/2022
+
+### Fixed
+- Compress email logs setting
+- Email duplication in email shots
+
+### Added
+- "View emails" button on email shots dashboard
+- Mailchimp lists integration through API
+
 ## 1.0.0 - 08/01/2022
+
+### Added
 - Initial release